Relatively happy with formatting now
Decided to switch to utilize eslint rather than tslint because tslint is going to be deprecated sometime in 2019 (so now?) in favor of using eslint in Javascript & Typescript.

One problem is that I use prettier to do formatting and there are certain things that bug me with prettier that are not options to configure. Not sure if there will ever be an answer around this.